refactor: refactored folder structure and rename functions
